requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
drf-yasg==1.20.0
psycopg2==2.9.1